Live 11 Beta build 22 fix
The latest Live 11 Beta build (b22) broke the script - tracked it down to an import that all the official scripts now have. Updated mine correspondingly and it works. You'll probably need to delete the _pycache_ dir after updating the script (this is created each time Live runs and 'compiles' the script to python bytecode)